minor adaption
