+++ include/ast.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <stdint.h>
-#include <assert.h>
 #include <stdexcept>
 #include <string>
 #include <vector>
